Move the Axiell-FOLIO sync mapping code out of transformers
The axiell_folio_sync package is not a transformer: transformers turn
Iceberg adapter records into SourceWork Elasticsearch documents, while
this code maps MARCXML to FOLIO Inventory payloads and upserts them via
OKAPI. It never participated in the transformer dispatch and its only
consumer is the adapters.steps.axiell_folio_sync step, so fold it into
that step's directory, making the service self-contained. Tests move to
mirror the new layout. The unused re-export facade __init__ is dropped;
all consumers import fully-qualified module paths.
