--- catalogue_graph/src/adapters/transformers/axiell_folio_sync/__init__.py
@@ -1,39 +1,0 @@
-"""Axiell to Folio sync prototype package."""
-
-from .mapper import CanonicalRecord
-from .mapping import (
-    EntityResult,
-    MappedPayloads,
-    MappingError,
-    PayloadMeta,
-    UpsertError,
-    UpsertResult,
-    build_holdings,
-    build_instance,
-    build_item,
-    build_payloads,
-    is_selected_for_sync,
-    parse_marcxml,
-    select_and_build,
-)
-from .ref_cache import RefCache
-from .upsert import upsert_from_payloads
-
-__all__ = [
-    "CanonicalRecord",
-    "EntityResult",
-    "MappedPayloads",
-    "MappingError",
-    "PayloadMeta",
-    "RefCache",
-    "UpsertError",
-    "UpsertResult",
-    "build_payloads",
-    "build_instance",
-    "build_holdings",
-    "build_item",
-    "is_selected_for_sync",
-    "parse_marcxml",
-    "select_and_build",
-    "upsert_from_payloads",
-]
